Add concurrency handling with a lock per mailer
If the lock command was run twice IOExceptions occur because both instances try to send the same email.

This happens for example if you run the command by cron every minute and you send a lot of emails. The first instance started by cron is still sending emails, when the next instance is started one minute later.